requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
absl-py=1.4.0=pypi_0
astor=0.8.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backpack-for-pytorch=1.5.2=pypi_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bitarray=2.7.3=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=39.0.1=py37h9ce1e76_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.1.55=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.1.55=0
cuda-nsight-compute=12.1.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.1.55=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.1.55=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.1.55=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.1.55=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
cycler=0.11.0=pypi_0
decorator=5.1.1=pyhd8ed1ab_0
einops=0.6.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
filelock=3.9.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
future=0.18.3=pypi_0
gast=0.5.3=pyhd3eb1b0_0
gds-tools=1.6.0.25=0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpcio=1.51.3=pypi_0
h5py=3.8.0=pypi_0
hdf5=1.10.6=h3ffc7dd_1
huggingface-hub=0.13.0=pypi_0
idna=3.4=py37h06a4308_0
importlib-metadata=6.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=5.5.5=py37h085eea5_0
ipython=7.33.0=py37h89c1867_0
ipython_genutils=0.2.0=py_1
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
jupyter_client=5.3.4=py37_0
jupyter_core=4.5.0=py_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.6.0.25=0
libcufile-dev=1.6.0.25=0
libcurand=10.3.2.56=0
libcurand-dev=10.3.2.56=0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py37h06a4308_0
markupsafe=2.1.2=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
mock=5.0.1=pypi_0
ncurses=6.4=h6a678d5_0
nearpy=1.0.0=pypi_0
nettle=3.7.3=hbbd107a_1
nsight-compute=2023.1.0.15=0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pypi_0
pandas=1.3.5=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py37h6a678d5_0
pip=22.3.1=py37h06a4308_0
prompt-toolkit=3.0.38=pyha770c72_0
protobuf=4.22.1=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.14.0=pyhd8ed1ab_0
pyopenssl=23.0.0=py37h06a4308_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py37_1
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-version=0.0.2=pypi_0
python_abi=3.7=2_cp37m
pytorch=1.13.1=py3.7_cuda11.6_cudnn8.3.2_0
pytorch-cinic=0.0.6=pypi_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py37hac76be4_2
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.1=py37h06a4308_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=py37h6c91a56_2
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h5082296_0
tensorboard=1.13.1=py37hf484d3e_0
tensorflow=1.13.2=pypi_0
tensorflow-base=1.13.1=mkl_py37h7ce6ba3_0
tensorflow-estimator=1.13.0=py_0
termcolor=2.2.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
torchaudio=0.13.1=py37_cu116
torchvision=0.14.1=py37_cu116
tornado=6.1=py37h540881e_3
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.26.1=pypi_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
warmup-scheduler=0.3=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.2.3=pypi_0
wget=3.2=pypi_0
wheel=0.38.4=py37h06a4308_0
xz=5.2.10=h5eee18b_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
